# docs/production/openapi.json
{
  "info": {
    "description": "Generated from the route table (server/build/routeTable.ts) and the request-contract registry (server/build/apiSurface.ts) by scripts/generate-openapi.ts. Request bodies are the zod schemas the handlers parse with, emitted by zod. Success response bodies are not yet described; x-coverage says how much of the surface each part of this document covers.",
    "title": "Abedin Growth AI API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/meetings/{id}/process-payment": {
      "post": {
        "operationId": "post_api_meetings_id_process_payment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/meetings.routes.ts:44"
      }
    },
    "/api/meetings/{id}/send-recovery-email": {
      "post": {
        "operationId": "post_api_meetings_id_send_recovery_email",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/meetings.routes.ts:54"
      }
    },
    "/api/meetings/{id}/send-reminder": {
      "post": {
        "operationId": "post_api_meetings_id_send_reminder",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/meetings.routes.ts:70"
      }
    },
    "/api/meetings/{id}/sign-contract": {
      "post": {
        "operationId": "post_api_meetings_id_sign_contract",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "ATTRIBUTION_REQUIRED",
                            "AUTH_INVALID",
                            "AUTH_REQUIRED",
                            "AUTH_UNAVAILABLE",
                            "CONFIGURATION_ERROR",
                            "CONTACT_EXISTS",
                            "ILLEGAL_TRANSITION",
                            "INTERNAL_ERROR",
                            "MERGE_REFUSED",
                            "MODEL_OUTPUT_INVALID",
                            "NOT_FOUND",
                            "NOT_IMPLEMENTED",
                            "POLICY_BLOCKED",
                            "PROVIDER_RATE_LIMITED",
                            "PROVIDER_UNAVAILABLE",
                            "STORE_UNAVAILABLE",
                            "TENANT_AMBIGUOUS",
                            "TENANT_FORBIDDEN",
                            "TENANT_INVALID",
                            "TENANT_REVOCATION_UNVERIFIABLE",
                            "TENANT_SUSPENDED",
                            "TENANT_UNRESOLVED",
                            "TERMINAL_STATE",
                            "TOO_MANY_REFERENCES",
                            "UNKNOWN_CURRENT_STATE",
                            "UNKNOWN_TARGET_STATE",
                            "VALIDATION_ERROR",
                            "VERSION_CONFLICT",
                            "VERSION_REQUIRED",
                            "WEBHOOK_VERIFICATION_FAILED"
                          ],
                          "type": "string"
                        },
                        "details": {
                          "type": "object"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Any error, in the envelope server/lib/errors.ts builds; the code is one of the taxonomy."
          }
        },
        "x-body": "none",
        "x-source": "server/routes/meetings.routes.ts:22"
      }
    },
...
  },
...
}